add FalconSecret spec to FalconDeployment CRD
make FalconAPI optional if FalconSecret is enabled

refactor falcon secret injection as a generic method

+++ internal/controller/admission/falconadmission_controller_test.go
@@ -7,6 +7,7 @@ import (
 
 	falconv1alpha1 "github.com/crowdstrike/falcon-operator/api/falcon/v1alpha1"
 	k8sutils "github.com/crowdstrike/falcon-operator/internal/controller/common"
+	internalErrors "github.com/crowdstrike/falcon-operator/internal/errors"
 	"github.com/crowdstrike/falcon-operator/pkg/common"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
@@ -114,23 +115,23 @@ var _ = Describe("FalconAdmission controller", func() {
 				return len(deployList.Items)
 			}, 6*time.Second, 2*time.Second).Should(Equal(0))
 
-			// Delete cluster level resources
-			clusterRoleBinding := &rbacv1.ClusterRoleBinding{}
-			Expect(k8sClient.Get(ctx, types.NamespacedName{Name: admissionClusterRoleBindingName}, clusterRoleBinding)).To(Succeed())
-			Expect(k8sClient.Delete(ctx, clusterRoleBinding)).To(Succeed())
-
-			if shouldAdmissionControlBeEnabled {
-				validatingWebhookConfig := &arv1.ValidatingWebhookConfiguration{}
-				Expect(k8sClient.Get(ctx, types.NamespacedName{Name: common.FalconAdmissionValidatingWebhookName}, validatingWebhookConfig)).To(Succeed())
-				Expect(k8sClient.Delete(ctx, validatingWebhookConfig)).To(Succeed())
-			}
-
 			// Delete FalconAdmission custom resource
 			falconAdmissionCR := &falconv1alpha1.FalconAdmission{}
 			Expect(k8sClient.Get(ctx, admissionNamespacedName, falconAdmissionCR)).To(Succeed())
 			Expect(k8sClient.Delete(ctx, falconAdmissionCR)).To(Succeed())
 
-			// TODO(user): Attention if you improve this code by adding other context test you MUST
+			// Delete cluster level resources
+			clusterRoleBinding := &rbacv1.ClusterRoleBinding{}
+			if err := k8sClient.Get(ctx, types.NamespacedName{Name: admissionClusterRoleBindingName}, clusterRoleBinding); err == nil {
+				Expect(k8sClient.Delete(ctx, clusterRoleBinding)).To(Succeed())
+			}
+
+			validatingWebhookConfig := &arv1.ValidatingWebhookConfiguration{}
+			err := k8sClient.Get(ctx, types.NamespacedName{Name: common.FalconAdmissionValidatingWebhookName}, validatingWebhookConfig)
+			if err == nil {
+				Expect(k8sClient.Delete(ctx, validatingWebhookConfig)).To(Succeed())
+			}
+
 			_ = k8sClient.Delete(ctx, namespace)
 		})
 
@@ -360,6 +361,68 @@ var _ = Describe("FalconAdmission controller", func() {
 				}
 				return nil
 			}, 20*time.Second, time.Second).Should(Succeed())
+
+			By("Cleaning up the test specific resources")
+			err = k8sClient.Delete(ctx, testSecret)
+			Expect(err).To(Not(HaveOccurred()))
+		})
+
+		It("should error if falcon secret is enabled and Falcon API credentials are missing", func() {
+			By("Creating test secrets - without Falcon API credentials")
+			secretName := "falcon-secrets-1"
+			testSecretNamespace := "falcon-secret-1"
+
+			falconSecretNamespace := corev1.Namespace{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      testSecretNamespace,
+					Namespace: testSecretNamespace,
+				},
+			}
+
+			err := k8sClient.Create(ctx, &falconSecretNamespace)
+			Expect(err).To(Not(HaveOccurred()))
+
+			testSecret := &corev1.Secret{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      secretName,
+					Namespace: testSecretNamespace,
+				},
+				Type: corev1.SecretTypeOpaque,
+				StringData: map[string]string{
+					"falcon-cid": falconCID,
+				},
+			}
+			err = k8sClient.Create(ctx, testSecret)
+			Expect(err).To(Not(HaveOccurred()))
+
+			By("Creating the FalconAdmission CR with FalconSecret configured")
+			falconAdmission.Spec.FalconSecret = falconv1alpha1.FalconSecret{
+				Enabled:    true,
+				Namespace:  testSecretNamespace,
+				SecretName: secretName,
+			}
+			falconAdmission.Spec.Image = ""
+
+			err = k8sClient.Create(ctx, falconAdmission)
+			Expect(err).To(Not(HaveOccurred()))
+
+			By("Checking if the custom resource was successfully created")
+			Eventually(func() error {
+				falconAdmission := &falconv1alpha1.FalconAdmission{}
+				return k8sClient.Get(ctx, admissionNamespacedName, falconAdmission)
+			}, 6*time.Second, time.Second).Should(Succeed())
+
+			By("Reconciling the custom resource created")
+			falconAdmissionReconciler := &FalconAdmissionReconciler{
+				Client: k8sClient,
+				Reader: k8sReader,
+				Scheme: k8sClient.Scheme(),
+			}
+
+			_, err = falconAdmissionReconciler.Reconcile(ctx, reconcile.Request{
+				NamespacedName: admissionNamespacedName,
+			})
+			Expect(err).To(MatchError(ContainSubstring(internalErrors.ErrMissingFalconAPICredentialsInSecret.Error())))
 
 			By("Cleaning up the test specific resources")
 			err = k8sClient.Delete(ctx, testSecret)
